✨ feature : home 페이지 컴포넌트 분리 및 css 분리
